engine: move targetqueue into buildcontrol
