+++ internal/engine/buildcontrol/target_queue_test.go
@@ -1,4 +1,4 @@
-package engine
+package buildcontrol
 
 import (
 	"context"
